make content more flexible
